Do not wrap git message for 'git tree command.'
Why is this necessary?
I want to properly see how branches are merged and the long commits
breaks the flow of the graph.

How does this address the problem?
Piping the result of 'git log' to 'more -S' prevents it from wrapping long
commit messages. The result is colored by passing --color=always to the
git command and -R to 'more' command to let it interpret those colors.

Are there any side effects?
Cannot see the whole commit message.